feature/travel-cron (#3)

+++ frontend/src/components/ListTravels.vue
@@ -1,20 +1,29 @@
 <template>
   <div class="bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4">
-    <p class="text-4xl mb-8">Travels</p>
+    <div class="flex justify-between items-center mb-8">
+      <p class="text-4xl">Travels</p>
+      <button
+        type="button"
+        class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none"
+        @click="showModal"
+      >
+        Add a travel
+      </button>
+    </div>
     <div v-if="this.$store.state.travels.length === 0" class="flex justify-center items-center">
       <img src="../assets/benoit.png" class="rounded-full">
     </div>
-    <div class="rounded overflow-hidden border-r border-b border-l border-t border-gray-400 mb-4" v-for="travel in this.$store.state.travels">
+    <div class="rounded overflow-hidden border-r border-b border-l border-t border-gray-400 mb-4" v-for="travel in this.$store.state.travels" v-bind:key="travel.id">
       <div class="px-6 py-4">
         <div class="font-bold text-xl">
           {{ travel.from }} → {{ travel.to }}
           <div class="text-gray-500 float-right text-base font-normal">
-            {{ new Date(travel.date).toISOString().split('T')[0] }}
-            <span v-if="travel.minHour"><span v-if="!travel.maxHour">></span>{{ travel.minHour }}h</span><span v-if="travel.minHour && travel.maxHour">-</span><span v-if="travel.maxHour"><span v-if="!travel.minHour"><</span>{{ travel.maxHour }}h</span>
+            {{ travel.date }}
+            <span v-if="travel.minHour"><span v-if="!travel.maxHour">></span>{{ travel.minHour }}h</span><span v-if="travel.minHour && travel.maxHour">-</span><span v-if="travel.maxHour"><span v-if="!travel.minHour"></span>{{ travel.maxHour }}h</span>
           </div>
         </div>
-        <p v-if="travel.book" class="text-gray-700 text-base">
-          Le train sera réservé avec {{ travel.booker.type }} ({{ travel.booker.name }})
+        <p class="text-gray-700 text-base">
+          Le train sera vérifié <span v-if="travel.book">et reservé </span>avec {{ travel.booker.type }} ({{ travel.booker.name }})
         </p>
         <p class="text-gray-700 text-base">
           Vous serez notifié via {{ travel.notifier.type }} ({{ travel.notifier.name }})
@@ -25,17 +34,42 @@
         <div class="clearfix"></div>
       </div>
     </div>
+
+    <ModalTravel
+      v-show="isModalVisible"
+      @close="closeModal"
+    />
   </div>
 </template>
 
 <script>
+
+import ModalTravel from './ModalTravel.vue'
+
 export default {
   name: "ListTravels",
+  components: {
+    ModalTravel
+  },
+  data() {
+    return {
+      isModalVisible: false,
+    };
+  },
   methods: {
     async deleteTravel (travel) {
+      if (travel.cron && !confirm(`Ce travel appartient a une Cron, il ne sera pas recréé s'il est supprimé. Etes-vous sur?`)) {
+        return
+      }
       await fetch(`${process.env.API_URL}/travels/${travel.id}`, { method: 'DELETE' })
       const index = this.$store.state.travels.findIndex(t => t.id === travel.id)
       this.$store.state.travels.splice(index, 1)
+    },
+    showModal() {
+      this.isModalVisible = true
+    },
+    closeModal() {
+      this.isModalVisible = false
     }
   }
 }
